requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-filter==24.2
djoser==2.2.3
